fix(dnd lists): refactored to reduce component size + memoized lists + sensors
